requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Computer Vision and Face Recognition
opencv-python==4.8.1.78
deepface==0.0.79
numpy>=1.26.0

pillow==10.1.0

# Barcode Scanning
pyzbar==0.1.9

# Data Processing
pandas==2.1.3

# PDF Generation
reportlab==4.0.7

# CORS is handled by FastAPI middleware
